refactor register hook
