+++ auto_round/algorithms/quantization/sign_round/quantizer.py
@@ -36,7 +36,6 @@ from auto_round.logger import logger
 from auto_round.utils import (
     check_to_quantized,
     compile_func,
-    convert_module_to_hp_if_necessary,
     get_module,
     htcore,
     is_auto_device_mapping,
@@ -44,7 +43,6 @@ from auto_round.utils import (
     memory_monitor,
     mv_module_from_gpu,
     set_amax_for_all_moe_layers,
-    set_module,
     to_device,
 )
 from auto_round.utils.device import (
@@ -345,28 +343,11 @@ class SignRoundQuantizer(BaseQuantizers):
         """
         if input_ids is None:
             logger.info(f"using rtn to quantize {layer_name}")
-            layer = get_module(self.model, layer_name)
-            if hasattr(layer, "tuning_device"):
-                device = layer.tuning_device
-            if dtype is not None:
-                layer = layer.to(dtype)
-            layer = layer.to(device)
-            layer = convert_module_to_hp_if_necessary(layer, self.model_context.amp_dtype, device)
-            set_module(self.model, layer_name, layer)
-
-            wrapper_linear = WrapperLinear(
-                layer,
-                enable_round_tuning=False,
-                enable_minmax_tuning=False,
-                enable_norm_bias_tuning=False,
-                enable_torch_compile=self.compress_context.enable_torch_compile,
-                device=device,
+            self.quantize_layer_via_rtn(
+                layer_name,
+                dtype=dtype,
                 disable_opt_rtn=getattr(self.config, "disable_opt_rtn", True),
-                iters=0,
             )
-            new_layer = wrapper_linear.unwrapper({})
-            set_module(self.model, layer_name, new_layer)
-            layer.cpu()
             return
 
         logger.info(f"quantizing layer {layer_name}")
@@ -390,7 +371,7 @@ class SignRoundQuantizer(BaseQuantizers):
             static_attention_dtype,
         ):
             tmp_inputs = q_inputs if q_inputs is not None else input_ids
-            hook_handles = self._register_act_max_hook(layer)
+            hook_handles = self.register_calibration_hooks(layer)
             with torch.no_grad():
                 for input in tmp_inputs:
                     layer(input)
